pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["Igor <iiigorrr@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.4"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
asyncio = "^3.4.3"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.2.0"
types-redis = "*"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pre-commit = "^4.0.1"
pylint = "^3.3.1"
types-redis = "^4.6.0.20241004"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
